Make Quoting Dojo Modularization
