+++ quoting_dojo/package.json
@@ -14,8 +14,9 @@
     "ejs": "^2.4.1",
     "express": "^4.16.2",
     "express-session": "^1.15.6",
+    "fs": "0.0.1-security",
+    "moment": "^2.20.1",
     "mongoose": "^5.0.4",
-    "socket.io": "^2.0.4",
-    "moment": "^2.20.1"
+    "socket.io": "^2.0.4"
   }
 }
