+++ quoting_dojo/server/config/mongoose.js
@@ -1,0 +1,14 @@
+let mongoose = require('mongoose');
+//require file system module for loading model files
+let fs = require('fs');
+let path = require("path");
+mongoose.connect('mongodb://localhost/basic_mongoose');
+mongoose.Promise = global.Promise;
+var models_path = path.join(__dirname, './../models');
+// read all of the files in the models_path and require (run) each of the javascript files
+fs.readdirSync(models_path).forEach(function(file) {
+  if(file.indexOf('.js') >= 0) {
+    // require the file (this runs the model file which registers the schema)
+    require(models_path + '/' + file);
+  }
+});
